requirements.txt
aiofiles==23.1.0
alembic==1.9.4
async-generator==1.10
asyncio==3.4.3
attrs==22.2.0
autopep8==2.0.1
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
click==8.1.3
colorama==0.4.6
exceptiongroup==1.1.0
Flask==2.2.3
Flask-Cors==3.0.10
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
fyers-apiv2==2.0.5
greenlet==2.0.2
h11==0.14.0
idna==2.10
importlib-metadata==6.0.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
outcome==1.2.0
psycopg2==2.9.5
PyAutoIt==0.6.5
pycodestyle==2.10.0
pycparser==2.21
pyotp==2.8.0
PySocks==1.7.1
python-dotenv==1.0.0
pywin32==305
requests==2.25.1
selenium==4.8.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
tornado==6.1
trio==0.22.0
trio-websocket==0.9.2
typing-extensions==4.5.0
urllib3==1.26.14
websocket-client==1.2.1
websockets==8.1
Werkzeug==2.2.3
wsproto==1.2.0
xlwings==0.29.1
zipp==3.14.0
